+++ service/consul/datacenterWatch_test.go
@@ -9,19 +9,21 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"github.com/xmidt-org/argus/chrysom"
 	"github.com/xmidt-org/argus/model"
 	"github.com/xmidt-org/webpa-common/service"
-	"github.com/xmidt-org/webpa-common/xmetrics/xmetricstest"
+	"github.com/xmidt-org/webpa-common/xmetrics"
 )
 
 func TestNewDatacenterWatcher(t *testing.T) {
 	logger := log.NewNopLogger()
-	p := xmetricstest.NewProvider(nil, chrysom.Metrics)
+	r, err := xmetrics.NewRegistry(nil, Metrics)
+	require.Nil(t, err)
 	envShutdownChan := make(<-chan struct{})
 
 	mockServiceEnvironment := new(service.MockEnvironment)
-	mockServiceEnvironment.On("Provider").Return(p, true)
+	mockServiceEnvironment.On("Provider").Return(r, true)
 	mockServiceEnvironment.On("Closed").Return(envShutdownChan)
 
 	noProviderEnv := new(service.MockEnvironment)
@@ -77,7 +79,7 @@ func TestNewDatacenterWatcher(t *testing.T) {
 				mockServiceEnvironment, new(mockClient),
 			},
 			options: Options{
-				ChrysomConfig: chrysom.ClientConfig{
+				Chrysom: chrysom.ClientConfig{
 					Bucket: "",
 				},
 			},
@@ -100,7 +102,7 @@ func TestNewDatacenterWatcher(t *testing.T) {
 				mockServiceEnvironment, new(mockClient),
 			},
 			options: Options{
-				ChrysomConfig: validChrysomConfig,
+				Chrysom: validChrysomConfig,
 			},
 			expectedWatcher: &datacenterWatcher{
 				logger: logger,
@@ -109,7 +111,7 @@ func TestNewDatacenterWatcher(t *testing.T) {
 				},
 				options: Options{
 					DatacenterWatchInterval: defaultWatchInterval,
-					ChrysomConfig:           validChrysomConfig,
+					Chrysom:                 validChrysomConfig,
 				},
 				consulWatchInterval: defaultWatchInterval,
 				inactiveDatacenters: make(map[string]bool),
@@ -124,7 +126,7 @@ func TestNewDatacenterWatcher(t *testing.T) {
 			},
 			options: Options{
 				DatacenterWatchInterval: 10 * time.Second,
-				ChrysomConfig:           validChrysomConfig,
+				Chrysom:                 validChrysomConfig,
 			},
 			expectedWatcher: &datacenterWatcher{
 				logger: logger,
@@ -133,7 +135,7 @@ func TestNewDatacenterWatcher(t *testing.T) {
 				},
 				options: Options{
 					DatacenterWatchInterval: 10 * time.Second,
-					ChrysomConfig:           validChrysomConfig,
+					Chrysom:                 validChrysomConfig,
 				},
 				inactiveDatacenters: make(map[string]bool),
 				consulWatchInterval: 10 * time.Second,
@@ -188,7 +190,7 @@ func TestNewDatacenterWatcher(t *testing.T) {
 				noProviderEnv, new(mockClient),
 			},
 			options: Options{
-				ChrysomConfig: validChrysomConfig,
+				Chrysom: validChrysomConfig,
 			},
 			expectedErr: errors.New("must pass in a metrics provider"),
 		},
@@ -199,7 +201,7 @@ func TestNewDatacenterWatcher(t *testing.T) {
 				mockServiceEnvironment, new(mockClient),
 			},
 			options: Options{
-				ChrysomConfig: chrysom.ClientConfig{
+				Chrysom: chrysom.ClientConfig{
 					Bucket: "random-bucket",
 					Listen: chrysom.ListenerConfig{
 						PullInterval: 0,